fix: remove duplicate getProductSuggestions from routes, use controller export (#165)
